# openApi.yaml
openapi: 3.0.0
info:
  title: ETH-HOUSE
  version: "1.0.0"
  description: Learning swagger in Zebpay

servers:
  - url: http://localhost:3000
tags: 
   - name: Books
     description: Book APIS
paths:
  /books:
    get:
      summary: Returns a list of books
      tags: 
        - Books
      responses:
        200:
          description: The List of all books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      summary: Create a new book
      tags: 
       - Books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
      responses:
        200:
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        500:
          description: server error
  /books/{id}:
    get:
      summary: Returns book with the given id
      tags: 
       - Books
      parameters:
        - in: path
          name: id
          schema:
          type: string
          required: true
          description: The book id
      responses:
        200:
          description: Book with given id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        404:
          description: Book not found
    put:
      summary: Update an existing book
      tags:
        - Books
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The book id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
      responses:
        200:
          description: book updated sucessfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        404:
          description: book not found
        500:
          description: server error
components:
  schemas:
    Book:
      type: object
      required:
        - title
        - author
      properties:
        id:
          type: string
          description: the auto generated id
        title:
          type: string
          descprition: title of book
        author:
          type: string
          description: author of book
      example:
        id: abc
        title: eth_Blockchain
        author: Raghav Gupta